refactor: 함수 분리 및 파일 구조 구성
