+++ index.html
@@ -5,12 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Jay Zoey Village</title>
-    <link rel="stylesheet" href="./style.css" />
-    <script type="module" src="./main.js" defer></script>
+    <link rel="stylesheet" href="./src/css/style.css" />
+    <script type="module" src="./src/js/main.js" defer></script>
   </head>
   <body>
     <header>
-      <h1>마을 지도</h1>
+      <h1 class="title">마을 지도</h1>
     </header>
     <main>
       <div class="root"></div>
