+++ src/js/utils/domAPI.js
@@ -1,4 +1,3 @@
-//DFSPreOrder
 export const domAPI = {
   htmlNode: document.firstElementChild,
