+++ src/js/view/makeVillageView.js
